added navigation base code
