+++ Mobile App/app/src/main/java/com/example/scar/MainActivity.kt
@@ -52,7 +52,7 @@ class MainActivity  : ComponentActivity() {
         setContent {
             SCARTheme {
                 // A surface container using the 'background' color from the theme
-                leaderboard()
+                Navigation()
 //
 //                val url = "http://192.168.1.5:5000//match/1"  // Replace with your actual URL
 //
@@ -79,157 +79,4 @@ class MainActivity  : ComponentActivity() {
             }
         }
     }
-}
-data class LeaderboardEntry(val name: String, val score: Int)
-
-@Composable
-fun leaderboard(modifier: Modifier = Modifier){
-    Surface(modifier = Modifier.fillMaxSize(),
-        color = mainBg// Set the desired background color here
-    ) {
-        Column {
-            Row(
-                horizontalArrangement = Arrangement.Center,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Button(
-                    onClick = {
-                        // Handle button click for the first button
-                    },
-                    shape = RectangleShape,
-                    colors = ButtonDefaults.buttonColors(
-                        cardBg.copy(alpha = 0.6f)
-                    ),
-                    modifier = Modifier
-                        .size(150.dp)
-                        .padding(15.dp)
-                        .clip(shape = RoundedCornerShape(6.dp))
-                        .border(3.dp, Color.White)
-                )
-                {
-                    Text(
-                        text = "Regional",
-                        style = TextStyle(color = Color.White)
-                    )
-                }
-                Button(
-                    onClick = {
-                        // Handle button click for the first button
-                    },
-                    shape = RectangleShape,
-                    colors = ButtonDefaults.buttonColors(
-                        cardBg.copy(alpha = 0.6f)
-                    ),
-                    modifier = Modifier
-                        .size(150.dp)
-                        .padding(15.dp)
-                        .clip(shape = RoundedCornerShape(6.dp))
-                )
-                {
-                    Text(text = "World")
-                }
-            }
-            Row {
-                LeaderboardUI()
-            }
-
-        }
-    }
-}
-
-@Composable
-fun LeaderboardScreen(leaderboardData: List<LeaderboardEntry>) {
-    var searchText by remember { mutableStateOf("") }
-    Column(
-        modifier = Modifier
-            .padding(16.dp)
-            .clip(shape = RoundedCornerShape(8.dp))
-            .background(cardBg.copy(alpha = 0.6f))
-
-        ,
-        horizontalAlignment = Alignment.CenterHorizontally,
-
-        ) {
-        Text(
-            text = "LEADERBOARDS",
-            style = TextStyle(color= Color.White, fontSize = 20.sp),
-        )
-
-
-        // Leaderboard entries
-        LazyColumn {
-            itemsIndexed(leaderboardData.filter {
-                it.name.contains(
-                    searchText,
-                    ignoreCase = true
-                )
-            }) { index, entry ->
-                LeaderboardItem(entry = entry, isGold = index < 3)
-            }
-        }
-        Spacer(modifier = Modifier.height(90.dp))
-        Row(){
-            LeaderboardItem(entry = LeaderboardEntry("player",123), isGold = false)
-
-        }
-
-    }
-
-
-}
-
-@SuppressLint("SuspiciousIndentation")
-@Composable
-fun LeaderboardItem(entry: LeaderboardEntry, isGold: Boolean) {
-    val backgroundColor = if (isGold) Color(0xFFFFD700) else Color.White  // Use gold color for the first three items
-    ElevatedCard (
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-            .clip(RoundedCornerShape(8)),
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = 8.dp
-        ),
-        colors = CardDefaults.cardColors(containerColor = backgroundColor)
-
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(text = entry.name, style = MaterialTheme.typography.bodyMedium)
-            Text(text = entry.score.toString(), style = MaterialTheme.typography.bodyMedium)
-        }
-
-    }
-
-}
-
-@Composable
-fun LeaderboardUI() {
-    val leaderboardData = remember {
-        listOf(
-            LeaderboardEntry("User 1", 500),
-            LeaderboardEntry("User 2", 750),
-            LeaderboardEntry("User 3", 600),
-            LeaderboardEntry("User 4", 900),
-            LeaderboardEntry("User 5", 800),
-            LeaderboardEntry("User 6", 700),
-            LeaderboardEntry("User 7", 550),
-        )
-    }
-
-    LeaderboardScreen(leaderboardData = leaderboardData)
-}
-
-
-@Preview(
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true,
-    name = "Dark Mode")
-@Composable
-fun GreetingPreview() {
-    leaderboard()
 }
