+++ Mobile App/app/src/main/java/com/example/scar/ui/theme/Leaderboard.kt
@@ -1,0 +1,6 @@
+package com.example.scar.ui.theme
+
+sealed class Leaderboard(val route:String){
+    object Regional : Leaderboard("regional_leaderboard")
+    object World : Leaderboard("world_leaderboard")
+}
