+++ Mobile App/app/src/main/java/com/example/scar/ui/theme/Screen.kt
@@ -1,0 +1,16 @@
+package com.example.scar.ui.theme
+
+sealed class Screen(val route:String){
+    object MainScreen : Screen("main_screen")
+    object DetailScreen : Screen("detail_screen")
+
+    fun withArgs(args:String): String{
+        return buildString {
+            append(route)
+            args.forEach {arg->
+                append("/$arg")
+            }
+        }
+    }
+
+}
